rename config folder
