+++ src/services/index.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import config from '../config.js';
+import config from '../config';
 
 axios.defaults.baseURL = config.api_domain;
